refactor(code): new function to Box object
